refactor(upgrade-tests): factor swingStore-lite out of vat-status
