+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-12/tools/swingStore-lite.mjs
@@ -1,0 +1,68 @@
+// @ts-check
+
+/**
+ * @file enough swingStore access to look up incarnation by vat name
+ * @see {makeSwingstore}
+ *
+ * This is an independent implementation of the swingstore in order
+ * to test what's actually in the DB.
+ */
+
+export const swingstorePath = '~/.agoric/data/agoric/swingstore.sqlite';
+
+/**
+ * SQL short-hand
+ *
+ * @param {import('better-sqlite3').Database} db
+ */
+export const dbTool = db => {
+  const prepare = (strings, ...params) => {
+    const dml = strings.join('?');
+    return { stmt: db.prepare(dml), params };
+  };
+  const sql = (strings, ...args) => {
+    const { stmt, params } = prepare(strings, ...args);
+    return stmt.all(...params);
+  };
+  sql.get = (strings, ...args) => {
+    const { stmt, params } = prepare(strings, ...args);
+    return stmt.get(...params);
+  };
+  return sql;
+};
+
+/**
+ * @param {import('better-sqlite3').Database} db
+ */
+export const makeSwingstore = db => {
+  const sql = dbTool(db);
+
+  /** @param {string} key */
+  const kvGet = key => sql.get`select * from kvStore where key = ${key}`.value;
+  /** @param {string} key */
+  const kvGetJSON = key => JSON.parse(kvGet(key));
+
+  /** @param {string} vatID */
+  const lookupVat = vatID => {
+    return Object.freeze({
+      source: () => kvGetJSON(`${vatID}.source`),
+      options: () => kvGetJSON(`${vatID}.options`),
+      currentSpan: () =>
+        sql.get`select * from transcriptSpans where isCurrent = 1 and vatID = ${vatID}`,
+    });
+  };
+
+  return Object.freeze({
+    /** @param {string} vatName */
+    findVat: vatName => {
+      /** @type {string[]} */
+      const dynamicIDs = kvGetJSON('vat.dynamicIDs');
+      const targetVat = dynamicIDs.find(
+        vatID => lookupVat(vatID).options().name === vatName,
+      );
+      if (!targetVat) throw Error(vatName);
+      return targetVat;
+    },
+    lookupVat,
+  });
+};
